brain/requirements.txt
mlx-lm
huggingface_hub
fastapi
uvicorn
